+++ src/Exceptions/SingularMatrix.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Tensor\Exceptions;
+
+class SingularMatrix extends RuntimeException
+{
+    //
+}